Create spListToTable.sql
This function should be run in the “SQL Commands” page in rock to create a stored function that the rest of the sql statements will use for their queries. This function will easily split options passed into a query string into a table that is searchable in an IN () statement. For example, we will pass Currency Type IDs in the query string like ‘157, 156, 20’ and this function will convert those strings of numbers into a table of Ints for a `Where ID IN()` type search.